requirements.txt
dash==4.1.0
dash-bootstrap-components==2.0.4
plotly==6.6.0
pandas==2.3.3
pyyaml==6.0.3
